Add support for lazy-loading by requiring `octokit/lazy`, rather than just `octokit`
