+++ lib/octokit/lazy.rb
@@ -1,0 +1,3 @@
+require 'octokit/default'
+require 'octokit/configurable'
+require 'octokit/base'
